configure: Use -Werror-implicit-function-declaration with GCC
Add -Werror-implicit-function-declaration to the default compiled
flags used with GCC. Add the option through a new variable ERRORFLAGS
instead of the existing WARNFLAGS to avoid using it in configure
tests. I think it's overall preferable not to fail tests because of
the warning - in some case this could cause a compilation failure
later, but on the other hand it could be just an unreliable test
triggering the warning and even if it does fail those are likely cases
worth closer investigation.