mplayer: remove ---help option
mplayer used to follow the convention of using a single dash ("-help") for
each option. For some reason "--help" was allowed as special case.

mplayer2 introduced a change to allow the more common convention with two
dashes ("--help"). This means that in combination with the old special
case for "--help", "---help" would suddenly be allowed.

Remove this special case.