ao_coreaudio: fix partial volume control
If digital pass-through is used, this supported setting the volume (just
mute, actually), but not getting the volume. This will probably lead to a
stuck mute state in the mplayer frontend. Make the code respond to volume
queries even if digital pass-through is used.

Ideally, ao_coreaudio should implement full mute control, but I can't
even test on OSX.